requirements.txt
# Core
torch>=2.3.0
torchvision>=0.18.0
torchaudio>=2.3.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Imaging / audio utils
matplotlib>=3.8.0
pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.4.0
librosa>=0.10.1
soundfile>=0.12.1

# Experiment tracking (optional)
wandb>=0.17.0

# Supabase client (Python)
supabase>=2.6.0
python-dotenv>=1.0.0

# CLI helpers
rich>=13.7.0
tqdm>=4.66.0
medmnist>=3.0.1

# Server and testing
fastapi>=0.104.0
uvicorn>=0.24.0
pytest>=7.4.3
httpx>=0.25.0  # For TestClient
pytest-asyncio>=0.21.1

